// src/api/ambulance-wl/models/User.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Appointment Scheduling Api
 * Medical Appointment Scheduling System
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: xpokorny@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @interface User
 */
export interface User {
    /**
     * Unique identifier of the user
     * @type {string}
     * @memberof User
     */
    id: string;
    /**
     * Full name of the user
     * @type {string}
     * @memberof User
     */
    name: string;
    /**
     * Role of the user
     * @type {string}
     * @memberof User
     */
    role: UserRoleEnum;
}


/**
 * @export
 */
export const UserRoleEnum = {
    Patient: 'patient',
    Doctor: 'doctor'
} as const;
export type UserRoleEnum = typeof UserRoleEnum[keyof typeof UserRoleEnum];


/**
 * Check if a given object implements the User interface.
 */
export function instanceOfUser(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "role" in value;

    return isInstance;
}

export function UserFromJSON(json: any): User {
    return UserFromJSONTyped(json);
}

export function UserFromJSONTyped(json: any): User {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'role': json['role'],
    };
}

export function UserToJSON(value?: User | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'role': value.role,
    };
}